# Performance improvements to SpriteBatch 