--- a/MonoGame.Framework/Graphics/SpriteBatchItem.cs
+++ b/MonoGame.Framework/Graphics/SpriteBatchItem.cs
@@ -23,34 +23,34 @@ namespace Microsoft.Xna.Framework.Graphics
             vertexBR = new VertexPositionColorTexture();            
 		}
 		
-		public void Set ( float x, float y, float dx, float dy, float w, float h, float sin, float cos, Color color, Vector2 texCoordTL, Vector2 texCoordBR, float depth )
+		public void Set ( float x, float y, float w, float h, float sin, float cos, Color color, Vector2 texCoordTL, Vector2 texCoordBR, float depth )
 		{
             // TODO, Should we be just assigning the Depth Value to Z?
             // According to http://blogs.msdn.com/b/shawnhar/archive/2011/01/12/spritebatch-billboards-in-a-3d-world.aspx
             // We do.
-			vertexTL.Position.X = x+dx*cos-dy*sin;
-            vertexTL.Position.Y = y+dx*sin+dy*cos;
+            vertexTL.Position.X = x;
+            vertexTL.Position.Y = y;
             vertexTL.Position.Z = depth;
             vertexTL.Color = color;
             vertexTL.TextureCoordinate.X = texCoordTL.X;
             vertexTL.TextureCoordinate.Y = texCoordTL.Y;
 
-			vertexTR.Position.X = x+(dx+w)*cos-dy*sin;
-            vertexTR.Position.Y = y+(dx+w)*sin+dy*cos;
+            vertexTR.Position.X = x +w*cos;
+            vertexTR.Position.Y = y +w*sin;
             vertexTR.Position.Z = depth;
             vertexTR.Color = color;
             vertexTR.TextureCoordinate.X = texCoordBR.X;
             vertexTR.TextureCoordinate.Y = texCoordTL.Y;
 
-			vertexBL.Position.X = x+dx*cos-(dy+h)*sin;
-            vertexBL.Position.Y = y+dx*sin+(dy+h)*cos;
+            vertexBL.Position.X = x -h*sin;
+            vertexBL.Position.Y = y +h*cos;
             vertexBL.Position.Z = depth;
             vertexBL.Color = color;
             vertexBL.TextureCoordinate.X = texCoordTL.X;
             vertexBL.TextureCoordinate.Y = texCoordBR.Y;
 
-			vertexBR.Position.X = x+(dx+w)*cos-(dy+h)*sin;
-            vertexBR.Position.Y = y+(dx+w)*sin+(dy+h)*cos;
+            vertexBR.Position.X = x +w*cos-h*sin;
+            vertexBR.Position.Y = y +w*sin+h*cos;
             vertexBR.Position.Z = depth;
             vertexBR.Color = color;
             vertexBR.TextureCoordinate.X = texCoordBR.X;