--- a/MonoGame.Framework/Graphics/SpriteBatchItem.cs
+++ b/MonoGame.Framework/Graphics/SpriteBatchItem.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Xna.Framework.Graphics
             vertexBR = new VertexPositionColorTexture();            
 		}
 		
-		public void Set ( float x, float y, float w, float h, float sin, float cos, Color color, ref Vector2 texCoordTL, ref Vector2 texCoordBR, float depth )
+		public void Set ( float x, float y, float w, float h, float sin, float cos, Color color, Vector2 texCoordTL, Vector2 texCoordBR, float depth )
 		{
             // TODO, Should we be just assigning the Depth Value to Z?
             // According to http://blogs.msdn.com/b/shawnhar/archive/2011/01/12/spritebatch-billboards-in-a-3d-world.aspx
@@ -55,7 +55,7 @@ namespace Microsoft.Xna.Framework.Graphics
             vertexBR.TextureCoordinate = texCoordBR;
 		}
 
-        public void Set(float x, float y, float w, float h, Color color, ref Vector2 texCoordTL, ref Vector2 texCoordBR, float depth)
+        public void Set(float x, float y, float w, float h, Color color, Vector2 texCoordTL, Vector2 texCoordBR, float depth)
         {
             vertexTL.Position.X = x;
             vertexTL.Position.Y = y;