-Skip Calculations when Origin == Zero.
-Add _texelSize in Texture2D to avoid divisions in SpriteBatch.
 Used in calculating  origin & textureCord from sourceRectangle.


Current develop (Baseline)
33.83ms (100%)
Avoid extra calculations when Origin = Zero
30.38ms (89.8%)
Add _texelSize in Texture2D. Replace DIVs with MULs.
29.34ms (86.7%)

*test*
```
 var amount = 100000;
            var screenSize = 800;
            var spriteSize = 64;
            Rectangle textRect = new Rectangle(0,0,64,64);
            for (int i = 0; i < amount; i++)
            {
                var dest = new Rectangle(random.Next(screenSize), random.Next(screenSize), spriteSize, spriteSize);
                spriteBatch.Draw(_texture, dest, textRect, Color.White, random.Next(360), Vector2.Zero, SpriteEffects.None, 1);                
            }
```