@nkast 2016-10-31
Add my 'standard' DrawString() test
```

            for (var i = 20; i < 480; i += 20)
                this.spriteBatch.DrawString(this.fpsCounter.spriteFont, "abcdefghijklmnopqrstuvwxyz1234567890 some words here on the end test amazing testing letters to the edge of the screen", new Vector2(0, i), Color.Black);            
```

Current develop (Baseline)
1694fps 0.48ms (100%)
This PR
1823fps 0.44ms (91.7%)