@Vodacek 2016-10-31
Why not make TexelSize public?