@nkast 2016-11-22
Pushed another commit to move origin rotation out of SpriteBatchItem..Set().
Some perf benefit there as well as we do the calculations once.