@MichaelDePiazzi 2016-12-27
@nkast - I've been messing around with optimising `SpriteBatch` a bit lately, and I came up with a few changes that you might like to include in your PR. Of course I'm happy to submit them myself also if you don't want to include them. But in that case I'll wait until this PR is merged to avoid conflicts.

https://github.com/MichaelDePiazzi/MonoGame/compare/finaldays...MichaelDePiazzi:fdSpriteBatchOpt?w=1

Summary of changes:

- Moved the logic from `SpriteBatch.DrawInternal` to `SpriteBatchItem.Set` _(goal is to make inlining in `SpriteBatch.Draw` methods simpler as per @tomspilman's proposal in #5347)_
- Updated `Draw` methods to call `SpriteBatchItem.Set` directly
- Changed `SpriteBatchItem.Set` to take texture coordinates instead of the source rectangle
- Changed `SpriteBatchItem.Set` to take large struct parameters (i.e. >32-bit) by reference
- Changed `SpriteFont` to pre-calculate glyph texture coordinates when the font is constructed **(this gave a quite good performance increase!)**
- Simplified rotation calculations in `SpriteBatchItem.Set`
- Started inlining some of the calculations in the `SpriteBatch.Draw` methods

And here's some benchmarks:

Calling `SpriteBatch.Draw` 10,000,000 times | Baseline | Optimised | Comparison
--- | --- | --- | ---
**Light**<br>(without source rectangle, origin, rotation, or flipping) | 800ms | 779ms | 2.6% less time<br>2.7% faster
**Heavy**<br>(with source rectangle, origin, rotation, and flipping) | 1715ms | 1644ms | 4.1% less time<br>4.3% faster

Calling `SpriteBatch.DrawString` 100,000 times | Baseline | Optimised | Comparison
--- | --- | --- | ---
**Light**<br>(without origin, rotation, or flipping) | 693ms | 471ms | 32.0% less time<br>47.1% faster
**Heavy**<br>(with origin, rotation, and flipping) | 1263ms | 1087ms | 13.9% less time<br>16.2% faster