@nkast 2016-12-27
@MichaelDePiazzi 
Proposal #5347 is about inlining and then simplifying each method. I am trying to move most unrelated calculations out of .Set() for this reason. 
By moving the code from DrawInternal into Set() we are back on square one. For example, DrawString(...) need to set SortKey and sin/cos once. Also most Draw(...) methods don't need rotation, it's best to keep the two overloads. 

Which Draw() did you use for testing? Passing position (Vector2) by ref will help with Draw(Texture2D texture, Vector2 position, ...) but in the case of Draw(Texture2D texture, Rectangle dest, ...) requires to create a new `Vector2`.

When I changed texCoordTL/texCoordBR to `ref` I saw a small improvement ~2%-1% (which is within the margin of error with modern OS, I killed almost all services and did several runs to ensure there was a difference.) Later I tested https://github.com/MonoGame/MonoGame/compare/develop...nkast:tncSpriteBatchDraw with this PR and found that for the light Draw() I had to initialize _texCoordTL/BR with Vector2.Zero/One or use a member _texCoordZero/_texCoordOne which I found kind of ugly.
Perhaps I should drop the `ref` commit from this PR and re-evaluate. 
We can first test your proposal to use [MethodImpl(MethodImplOptions.AggressiveInlining)] on `.Set()` methods.
There's an oportunity to set directly `vertexTL/BR.TextureCoordinate = texCoordTL/BR` in two out of 4 cases, which some Jit frameworks might optimize better that `vertexTL/BR.TextureCoordinate.X = texCoordTL/BR.X; vertexTL/BR.TextureCoordinate.Y = texCoordTL/BR.Y` and we can't do that if texCoord is a `Vector4`.
For example commit https://github.com/MonoGame/MonoGame/pull/5267/commits/5784e8c7e9e41616048343b341c7c10019d4e100 was a bit faster.

> Changed SpriteFont to pre-calculate glyph texture coordinates when the font is constructed (this gave a quite good performance increase!)

That's impresive. I think we should apply this right after we inline DrawInternal() into DrawInto(), it's a clear win in performance.
After that we can simplify DrawInto() and finally inline it into DrawString().