@MichaelDePiazzi 2016-12-27
> By moving the code from DrawInternal into Set() we are back on square one.

Yes, I see what you mean. I have gone down the wrong path with this.

> Which Draw() did you use for testing?

I used `Draw(texture, position, sourceRectangle, color, rotation, origin, scale, effects, layerDepth)`. And you are right, the overload with the `destinationRectangle` parameter required creating of 2 new `Vector2`'s actually (but it currently is creating a new `Vector4` anyway).

> Perhaps I should drop the ref commit from this PR and re-evaluate.

I also wondered if passing by ref was worthwhile. The performance difference was negligible (on a PC at least) and in some places it required me to copy values into local variables so I could pass them (which then negated any performance benefit).

> We can first test your proposal to use [MethodImpl(MethodImplOptions.AggressiveInlining)] on .Set() methods.

I've messed with this a little, but I haven't had much luck with it giving me a noticeable performance increase yet. I'll try spend some more time on it later though if I can.

> That's impresive. I think we should apply this right after we inline DrawInternal() into DrawInto(), it's a clear win in performance.

Cool, sounds good.