@nkast 2016-12-27
> (but it currently is creating a new Vector4 anyway).

In DrawInternal() ? We can remove that on inlined methods.

> I've messed with this a little, but I haven't had much luck with it giving me a noticeable performance increase yet. I'll try spend some more time on it later though if I can.

I tested  `[MethodImpl(MethodImplOptions.AggressiveInlining)]` with `.Set()` on top of https://github.com/MonoGame/MonoGame/compare/develop...nkast:tncSpriteBatchDraw and got a 12.9% improvement! If it transfers well to mono, Native.Net, etc we might not need `ref` after all. We can also use it to refactor common code like `SortKey`. And also prefix with it small math methods in Vector2/3/4, Matrix, etc.

I will leave it to you if you want to post a PR.