@nkast 2016-12-27
@tomspilman , @KonajuGames Any thoughts on this PR? 
We could use the optimizations around _texelSize before moving forward with #5347