@MichaelDePiazzi 2016-12-27
> and got a 12.9% improvement!

That is excellent! I hadn't tried using it with `Set`, so that's great to hear that it can make a worthwhile difference!

> I will leave it to you if you want to post a PR.

Honestly, I won't have much more time to spend on this. So if you have a chance to work on it, then please feel free to submit a PR. Otherwise I'll get to it when I can.