@nkast 2016-12-27
> Honestly, I won't have much more time to spend on this. So if you have a chance to work on it, then please feel free to submit a PR. Otherwise I'll get to it when I can.

Cool, No problem.