@nkast 2017-02-16
Drop the commit about Texture2D.TexelSize as it is now addressed by #5453.
I still need to rebase this PR and see which changes are still relevant after #5401.